Add background to shadow & change shadows to em.
Adding a background to the shadow improves the way things look while the layer images are loading. Meanwhile, using em for the shadow helps the shadow scale while the container's font size scales.

I could see not using em for the shadow dimensions, but I really think the shadow should have a background.